requirements.txt
# 核心依赖
Flask==2.2.3
openai==0.27.0
requests==2.28.2

# 其他依赖
gunicorn==20.1.0
python-dotenv==1.0.0
